Support require:true for model.destroy
Closes #617, Closes #618
