+++ lib/base/model.js
@@ -5,6 +5,7 @@ var Backbone = require('backbone');
 
 var Events   = require('./events');
 var Promise  = require('./promise');
+var Errors   = require('../errors');
 
 // A list of properties that are omitted from the `Backbone.Model.prototype`, to create
 // a generic model base.
@@ -163,23 +164,7 @@ _.extend(ModelBase.prototype, _.omit(Backbone.Model.prototype, modelOmitted), Ev
 
   save: function() {},
 
-  // Destroy a model, calling a "delete" based on its `idAttribute`.
-  // A "destroying" and "destroyed" are triggered on the model before
-  // and after the model is destroyed, respectively. If an error is thrown
-  // during the "destroying" event, the model will not be destroyed.
-  destroy: Promise.method(function(options) {
-    options = options ? _.clone(options) : {};
-    var sync = this.sync(options);
-    options.query = sync.query;
-    return Promise.bind(this).then(function() {
-      return this.triggerThen('destroying', this, options);
-    }).then(function() {
-      return sync.del();
-    }).then(function(resp) {
-      this.clear();
-      return this.triggerThen('destroyed', this, resp, options);
-    }).then(this._reset);
-  })
+  destroy: function() {}
 
 });
