+++ lib/errors.js
@@ -10,6 +10,9 @@ module.exports = {
   EmptyError: createError('EmptyError'),
 
   // Thrown when an update affects no rows and {require: true} is passed in model.save.
-  NoRowsUpdatedError: createError('NoRowsUpdatedError')
+  NoRowsUpdatedError: createError('NoRowsUpdatedError'),
+
+  // Thrown when a delete affects no rows and {require: true} is passed in model.destroy.
+  NoRowsDeletedError: createError('NoRowsDeletedError')
 
 };
