+++ lib/model.js
@@ -241,6 +241,27 @@ _.extend(BookshelfModel.prototype, {
     .return(this);
   }),
 
+  // Destroy a model, calling a "delete" based on its `idAttribute`.
+  // A "destroying" and "destroyed" are triggered on the model before
+  // and after the model is destroyed, respectively. If an error is thrown
+  // during the "destroying" event, the model will not be destroyed.
+  destroy: Promise.method(function(options) {
+    options = options ? _.clone(options) : {};
+    var sync = this.sync(options);
+    options.query = sync.query;
+    return Promise.bind(this).then(function() {
+      return this.triggerThen('destroying', this, options);
+    }).then(function() {
+      return sync.del();
+    }).then(function(resp) {
+      if (options.require && resp === 0) {
+        throw new Errors.NoRowsDeletedError('EmptyResponse');
+      }
+      this.clear();
+      return this.triggerThen('destroyed', this, resp, options);
+    }).then(this._reset);
+  }),
+
   // Reset the query builder, called internally
   // each time a query is run.
   resetQuery: function() {
