+++ test/integration/model.js
@@ -528,6 +528,11 @@ module.exports = function(bookshelf) {
         return m.destroy();
       });
 
+      it('will throw an error when trying to destroy a non-existent object with {require: true}', function() {
+        return expect(new Site({id: 1337}).destroy({require: true}))
+          .to.be.rejectedWith(bookshelf.NoRowsDeletedError);
+      });
+
     });
 
     describe('resetQuery', function() {
